requirements.txt
openai
requests
PyYAML
schedule
